requirements.txt
pyyaml
requests
pydantic